Version two of this project was completed as a single commit for my resubmission.
Project was reviewed by Udacity Coach, Andrew, for the Front-End Nanodegree.
All code included in this submission represents a complete independent effort
where helper code was referenced and abstracted to overcome areas in the project
that were unfamiliar relative to my current ability level.

All helper code referenced for this project is included with proper attribution both
in the code files where the specific helper code abstraction was utilized, and again
at the end of my README file for Version Two.